dwc3_document
Introduction
The Synopsys DesignWare Core SuperSpeed USB 3.0 Controller (hereinafter referred to as DWC3) is a USB SuperSpeed compliant controller 
which can be configured in one of 4 ways:
  Peripheral-only configuration
  Host-only configuration
  Dual-Role configuration
  Hub configuration
Linux currently supports several versions of this controller. In all likelyhood, the version in your SoC is already supported. 
At the time of this writing, known tested versions range from 2.02a to 3.10a. As a rule of thumb, anything above 2.02a should work 
reliably well.